+++ cropclassification/preprocess/_prepare_input_BEFL.py
@@ -6,9 +6,8 @@ parcel that don't have a clear classification in the input file get class 'UNKNO
 
 import logging
 from pathlib import Path
-from typing import Optional
 
-import geofileops as gfo
+import pandas as pd
 from typing_extensions import deprecated
 
 import cropclassification.helpers.config_helper as conf
@@ -60,8 +59,7 @@ ndvi_latecrop_median = "latecrop_ndvi_median"
 # TODO: should be in REFE instead of hardcoded!!!
 late_main_crops = {
     "MON_CG_BIETEN": ["71", "91", "8532", "9532"],
-    "MON_CG_MAIS": ["201", "202"],
-    "IGNORE:LATE_MAINCROP": ["396"],
+    "MON_CG_MAIS_OF_MENGTEELT": ["201", "202", "396"],
 }
 
 
@@ -70,8 +68,7 @@ def prepare_input(
     classtype_to_prepare: str,
     classes_refe_path: Path,
     min_parcels_in_class: int,
-    output_dir: Path,
-):
+) -> pd.DataFrame:
     """Creates a file for use in the marker_cropclass functionality.
 
     It should be a csv file with the following columns:
@@ -87,8 +84,8 @@ def prepare_input(
 
     # Read input file
     logger.info(f"Read parceldata from {input_parcel_path}")
-    parceldata_df = gfo.read_file(input_parcel_path)
-    logger.info(f"Read Parceldata ready, info(): {parceldata_df.info()}")
+    parceldata_df = pdh.read_file(input_parcel_path)
+    logger.info(f"Read Parceldata ready, {len(parceldata_df)=}")
 
     # Check if the id column is present...
     if conf.columns["id"] not in parceldata_df.columns:
@@ -302,48 +299,32 @@ def prepare_input(
             min_parcels_in_class=min_parcels_in_class,
             is_groundtruth=True,
         )
-    elif classtype_to_prepare == "POPULAR-CROP":
-        parceldata_df = prepare_input_most_popular_crop(
-            parceldata_df=parceldata_df,
-            column_BEFL_cropcode=column_BEFL_crop_declared,
-            column_output_class=conf.columns["class_declared"],
-            classes_refe_path=classes_refe_path,
-            min_parcels_in_class=1,
-            is_groundtruth=False,
-        )
-        parceldata_df = prepare_input_most_popular_crop(
-            parceldata_df=parceldata_df,
-            column_BEFL_cropcode=column_BEFL_crop,
-            column_output_class=conf.columns["class"],
-            classes_refe_path=classes_refe_path,
-            min_parcels_in_class=min_parcels_in_class,
-            is_groundtruth=False,
-        )
-    elif classtype_to_prepare == "POPULAR-CROP-GROUNDTRUTH":
-        parceldata_df = prepare_input_most_popular_crop(
-            parceldata_df=parceldata_df,
-            column_BEFL_cropcode=column_BEFL_crop_gt_verified,
-            column_output_class=conf.columns["class_groundtruth"],
-            classes_refe_path=classes_refe_path,
-            min_parcels_in_class=min_parcels_in_class,
-            is_groundtruth=True,
-        )
     elif classtype_to_prepare.startswith("LATECROP"):
         if classtype_to_prepare.startswith("LATECROP-EARLY"):
             scope = "EARLY_MAINCROP"
         elif classtype_to_prepare.startswith("LATECROP-LATE"):
             scope = "LATE_MAINCROP"
-        else:
+        elif classtype_to_prepare.startswith("LATECROP-ALL"):
             scope = "ALL"
+        else:
+            raise ValueError(f"Unknown late crop scope: {classtype_to_prepare}")
 
         is_groundtruth = classtype_to_prepare.endswith("-GROUNDTRUTH")
+        if not is_groundtruth:
+            column_latecrop = column_BEFL_latecrop
+            column_latecrop2 = column_BEFL_latecrop2
+            column_output_class = conf.columns["class"]
+        else:
+            column_latecrop = column_BEFL_latecrop_gt_verified
+            column_latecrop2 = None
+            column_output_class = conf.columns["class_groundtruth"]
 
         parceldata_df = prepare_input_latecrop(
             parceldata_df=parceldata_df,
-            column_BEFL_latecrop=column_BEFL_latecrop,
-            column_BEFL_latecrop2=column_BEFL_latecrop2,
+            column_BEFL_latecrop=column_latecrop,
+            column_BEFL_latecrop2=column_latecrop2,
             column_BEFL_maincrop=column_BEFL_crop,
-            column_output_class=conf.columns["class"],
+            column_output_class=column_output_class,
             classes_refe_path=classes_refe_path,
             min_parcels_in_class=min_parcels_in_class,
             is_groundtruth=is_groundtruth,
@@ -438,13 +419,13 @@ def prepare_input(
 
 
 def prepare_input_cropgroup(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
     min_parcels_in_class: int,
     is_groundtruth: bool,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the cropgroup marker.
 
     It should be a csv file with the following columns:
@@ -641,11 +622,11 @@ def prepare_input_cropgroup(
 
 
 def prepare_input_croprotation(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the croprotation marker.
 
     It should be a csv file with the following columns:
@@ -778,11 +759,11 @@ def prepare_input_croprotation(
 
 
 def prepare_input_carbonsupply(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the carbonsupply marker.
 
     It should be a csv file with the following columns:
@@ -916,13 +897,13 @@ def prepare_input_carbonsupply(
 
 @deprecated("fabacae is deprecated, add refe before using")
 def prepare_input_fabaceae(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
     min_parcels_in_class: int,
     is_groundtruth: bool,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the fabaceae marker.
 
     It should be a csv file with the following columns:
@@ -1124,16 +1105,16 @@ def prepare_input_fabaceae(
 
 
 def prepare_input_latecrop(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_latecrop: str,
-    column_BEFL_latecrop2: Optional[str],
+    column_BEFL_latecrop2: str | None,
     column_BEFL_maincrop: str,
     column_output_class: str,
     classes_refe_path: Path,
     min_parcels_in_class: int,
     is_groundtruth: bool,
     scope: str,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the latecrop marker.
 
     It should be a csv file with the following columns:
@@ -1254,8 +1235,8 @@ def prepare_input_latecrop(
     # If limited scope, apply it
     if scope == "EARLY_MAINCROP":
         # Only process parcels with an early main crop.
-        # Hence, set parcels with a main crop that stays late on the field to an IGNORE
-        # class.
+        # Hence, set parcels with a main crop that stays late on the field to a class
+        # that can be ignored later on.
         for class_name, cropcodes in late_main_crops.items():
             parceldata_df.loc[
                 parceldata_df[column_BEFL_maincrop].isin(cropcodes), column_output_class
@@ -1360,10 +1341,33 @@ def prepare_input_latecrop(
                     column_output_class,
                 ] = "IGNORE:NOT_ENOUGH_SAMPLES"
 
-        # Add copy of class as class_declared
-        parceldata_df[conf.columns["class_declared"]] = parceldata_df[
-            column_output_class
-        ]
+    # Add copy of class as class_declared
+    parceldata_df[conf.columns["class_declared"]] = parceldata_df[column_output_class]
+
+    # Add a column with a correction factor to use when applying doubt thresholds.
+    # Determine the correction factor based on the EOC score at the end of
+    # class_declared.
+    def get_eoc_score(class_declared: str) -> float:
+        """Get the EOC score based on the declared class."""
+        # If there is an EOC score, it is the last value in a "_" separated string
+        class_declared_parts = class_declared.split("_")
+        eoc_score_str = class_declared_parts[-1]
+        try:
+            eoc_score = float(eoc_score_str)
+        except ValueError:
+            eoc_score = 0.0
+
+        return eoc_score
+
+    parceldata_df["eoc_score"] = parceldata_df[conf.columns["class_declared"]].apply(
+        lambda x: get_eoc_score(x)
+    )
+    eoc_score_max = parceldata_df["eoc_score"].max()
+    if eoc_score_max == 0:
+        raise ValueError("Maximum EOC score is 0, cannot calculate correction factor")
+    parceldata_df["proba_correction_factor"] = (
+        parceldata_df["eoc_score"] / eoc_score_max
+    )
 
     """
     # Add IGNORE:FOR_TRAINING column: if 1, ignore for training
@@ -1420,6 +1424,7 @@ def prepare_input_latecrop(
                 ndvi_latecrop_count,
                 ndvi_latecrop_median,
                 "IGNORE:FOR_TRAINING",
+                "proba_correction_factor",
             ]
             and column not in conf.preprocess.getlist("extra_export_columns")
             and column not in columns_BEFL_to_keep
@@ -1434,13 +1439,13 @@ def prepare_input_latecrop(
 
 
 def prepare_input_cropgroup_early(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
     min_parcels_in_class: int,
     is_groundtruth: bool,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the cropgroup_early marker."""
     # First run the standard landcover prepare
     parceldata_df = prepare_input_cropgroup(
@@ -1477,11 +1482,11 @@ def prepare_input_cropgroup_early(
 
 
 def prepare_input_croprotation_early(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the croprotation_early marker."""
     # First run the standard landcover prepare
     parceldata_df = prepare_input_croprotation(
@@ -1513,11 +1518,11 @@ def prepare_input_croprotation_early(
 
 
 def prepare_input_carbonsupply_early(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the carbonsupply_early marker."""
     # First run the standard landcover prepare
     parceldata_df = prepare_input_carbonsupply(
@@ -1549,13 +1554,13 @@ def prepare_input_carbonsupply_early(
 
 
 def prepare_input_landcover(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
     min_parcels_in_class: int,
     is_groundtruth: bool,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the landcover marker.
 
     It should be a csv file with the following columns:
@@ -1700,13 +1705,13 @@ def prepare_input_landcover(
 
 
 def prepare_input_landcover_early(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
     min_parcels_in_class: int,
     is_groundtruth: bool,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the landcover_early marker."""
     # First run the standard landcover prepare
     parceldata_df = prepare_input_landcover(
@@ -1744,13 +1749,13 @@ def prepare_input_landcover_early(
 
 
 def prepare_input_ruggenteelt(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
     min_parcels_in_class: int,
     is_groundtruth: bool,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the ruggenteelt marker.
 
     It should be a csv file with the following columns:
@@ -1895,13 +1900,13 @@ def prepare_input_ruggenteelt(
 
 
 def prepare_input_ruggenteelt_early(
-    parceldata_df,
+    parceldata_df: pd.DataFrame,
     column_BEFL_cropcode: str,
     column_output_class: str,
     classes_refe_path: Path,
     min_parcels_in_class: int,
     is_groundtruth: bool,
-):
+) -> pd.DataFrame:
     """Prepare input file for use in the ruggenteelt_early marker."""
     # First run the standard ruggenteelt prepare
     parceldata_df = prepare_input_ruggenteelt(
@@ -1935,75 +1940,4 @@ def prepare_input_ruggenteelt_early(
             "new grassland cannot be ignored!"
         )
 
-    return parceldata_df
-
-
-@deprecated("most populair crop is deprecated, add refe before using")
-def prepare_input_most_popular_crop(
-    parceldata_df,
-    column_BEFL_cropcode: str,
-    column_output_class: str,
-    classes_refe_path: Path,
-    min_parcels_in_class: int,
-    is_groundtruth: bool,
-):
-    """Prepare input file for use in the most_popular_crop marker.
-
-    It should be a csv file with the following columns:
-        - object_id: column with a unique identifier
-        - classname: a string column with a readable name of the classes that will be
-          classified to
-
-    This specific implementation converts the typical export format used in BE-Flanders
-    to this format.
-    """
-    # Add columns for the class to use...
-    parceldata_df.insert(0, column_output_class, None)
-
-    parceldata_df.loc[
-        parceldata_df[column_BEFL_cropcode].isin(["60", "700", "3432"]),
-        column_output_class,
-    ] = "Grassland"  # Grassland
-    parceldata_df.loc[
-        parceldata_df[column_BEFL_cropcode].isin(["201", "202"]), column_output_class
-    ] = "Maize"  # Maize
-    parceldata_df.loc[
-        parceldata_df[column_BEFL_cropcode].isin(["901", "904"]), column_output_class
-    ] = "Potatoes"  # Potatoes
-    parceldata_df.loc[
-        parceldata_df[column_BEFL_cropcode].isin(["311", "36"]), column_output_class
-    ] = "WinterWheat"  # Winter wheat of spelt
-    parceldata_df.loc[
-        parceldata_df[column_BEFL_cropcode].isin(["91"]), column_output_class
-    ] = "SugarBeat"  # Sugar beat
-    parceldata_df.loc[
-        parceldata_df[column_BEFL_cropcode].isin(["321"]), column_output_class
-    ] = "WinterBarley"  # Winter barley
-    parceldata_df.loc[
-        parceldata_df[column_BEFL_cropcode].isin(["71"]), column_output_class
-    ] = "FodderBeat"  # Fodder beat
-
-    # Some extra cleanup: classes starting with 'nvt' or empty ones
-    logger.info(
-        "Set classes that are empty to 'IGNORE:UNIMPORTANT_CLASS' so they are ignored "
-        "further on..."
-    )
-    parceldata_df.loc[
-        parceldata_df[column_output_class].isnull(), column_output_class
-    ] = "IGNORE:UNIMPORTANT_CLASS"
-
-    # Drop the columns that aren't useful at all
-    for column in parceldata_df.columns:
-        if (
-            column not in [conf.columns["id"], column_output_class]
-            and column not in conf.preprocess.getlist("extra_export_columns")
-            and not column == column_BEFL_cropcode
-        ):
-            parceldata_df.drop(column, axis=1, inplace=True)
-        elif column == column_BEFL_gesp_pm:
-            parceldata_df[column_BEFL_gesp_pm] = parceldata_df[
-                column_BEFL_gesp_pm
-            ].str.replace(",", ";")
-
-    # Return result
     return parceldata_df