+++ cropclassification/taskrunner.py
@@ -8,17 +8,19 @@ from pathlib import Path
 # Import geofilops here already, if tensorflow is loaded first leads to dll load errors
 import geofileops as gfo  # noqa: F401
 
+from cropclassification import calc_cover, calc_cropclass, calc_periodic_mosaic
 
-def _get_version():
+
+def _get_version() -> str:
     version_path = Path(__file__).resolve().parent / "version.txt"
-    with open(version_path) as file:
+    with version_path.open() as file:
         return file.readline()
 
 
 __version__ = _get_version()
 
 
-def main():
+def main() -> None:
     """Parse the arguments and run the tasks."""
     # Interprete arguments
     parser = argparse.ArgumentParser(add_help=False)
@@ -56,7 +58,7 @@ def main():
             sys.exit(1)
 
 
-def run_tasks(tasksdir: Path, config_overrules: list[str] = []):
+def run_tasks(tasksdir: Path, config_overrules: list[str] | None = None) -> None:
     """Run the tasks in the tasks directory.
 
     Args:
@@ -65,6 +67,9 @@ def run_tasks(tasksdir: Path, config_overrules: list[str] = []):
             configuration. They should be specified as a list of
             "<section>.<parameter>=<value>" strings. Defaults to [].
     """
+    if config_overrules is None:
+        config_overrules = []
+
     # Get the tasks and treat them
     task_paths = sorted(tasksdir.glob("task_*.ini"))
     for task_path in task_paths:
@@ -108,24 +113,18 @@ def run_tasks(tasksdir: Path, config_overrules: list[str] = []):
                 config_paths.append(Path(config_file_formatted))
 
         if action in ("calc_marker", "calc_cropclass"):
-            from cropclassification import calc_cropclass
-
             calc_cropclass.run_cropclass(
                 config_paths=config_paths,
                 default_basedir=default_basedir,
                 config_overrules=config_overrules,
             )
         elif action == "calc_cover":
-            from cropclassification import calc_cover
-
             calc_cover.run_cover(
                 config_paths=config_paths,
                 default_basedir=default_basedir,
                 config_overrules=config_overrules,
             )
         elif action == "calc_periodic_mosaic":
-            from cropclassification import calc_periodic_mosaic
-
             calc_periodic_mosaic.calc_periodic_mosaic_task(
                 config_paths=config_paths, default_basedir=default_basedir
             )