+++ cropclassification/util/mosaic_util.py
@@ -6,9 +6,10 @@ from copy import deepcopy
 from dataclasses import dataclass
 from datetime import datetime, timedelta
 from pathlib import Path
-from typing import Any, Optional
+from typing import Any
 
 import pyproj
+from pyproj import CRS
 
 from . import date_util, openeo_util, raster_index_util, raster_util
 
@@ -23,17 +24,17 @@ class ImageProfile:
     name: str
     satellite: str
     image_source: str
-    bands: Optional[list[str]] = None
-    collection: Optional[str] = None
-    time_reducer: Optional[str] = None
-    period_name: Optional[str] = None
-    period_days: Optional[int] = None
-    base_imageprofile: Optional[str] = None
-    index_type: Optional[str] = None
-    pixel_type: Optional[str] = None
-    max_cloud_cover: Optional[float] = None
-    process_options: Optional[dict] = None
-    job_options: Optional[dict] = None
+    bands: list[str] | None = None
+    collection: str | None = None
+    time_reducer: str | None = None
+    period_name: str | None = None
+    period_days: int | None = None
+    base_imageprofile: str | None = None
+    index_type: str | None = None
+    pixel_type: str | None = None
+    max_cloud_cover: float | None = None
+    process_options: dict | None = None
+    job_options: dict | None = None
 
     def __init__(
         self,
@@ -41,17 +42,17 @@ class ImageProfile:
         satellite: str,
         image_source: str,
         bands: list[str],
-        collection: Optional[str] = None,
-        time_reducer: Optional[str] = None,
-        period_name: Optional[str] = None,
-        period_days: Optional[int] = None,
-        base_image_profile: Optional[str] = None,
-        index_type: Optional[str] = None,
-        pixel_type: Optional[str] = None,
-        max_cloud_cover: Optional[float] = None,
-        process_options: Optional[dict] = None,
-        job_options: Optional[dict] = None,
-    ):
+        collection: str | None = None,
+        time_reducer: str | None = None,
+        period_name: str | None = None,
+        period_days: int | None = None,
+        base_image_profile: str | None = None,
+        index_type: str | None = None,
+        pixel_type: str | None = None,
+        max_cloud_cover: float | None = None,
+        process_options: dict | None = None,
+        job_options: dict | None = None,
+    ) -> None:
         """Profile of the image to be calculated.
 
         The `period_name` and `period_days` are used to determine the periods of the
@@ -153,7 +154,7 @@ class ImageProfile:
 
 def calc_periodic_mosaic(
     roi_bounds: tuple[float, float, float, float],
-    roi_crs: Optional[Any],
+    roi_crs: CRS | None,
     start_date: datetime,
     end_date: datetime,
     imageprofiles_to_get: list[str],
@@ -252,7 +253,7 @@ def calc_periodic_mosaic(
                 )
 
     # First get all mosaic images from openeo
-    _ = openeo_util.get_images(
+    openeo_util.get_images(
         images_from_openeo,
         delete_existing_openeo_jobs=delete_existing_openeo_jobs,
         raise_errors="raise" in on_missing_image,
@@ -298,8 +299,8 @@ def _is_image_outdated(image, images_available_delay: Optional[int] = None) -> b
 def _prepare_periods(
     start_date: datetime,
     end_date: datetime,
-    period_name: Optional[str],
-    period_days: Optional[int],
+    period_name: str | None,
+    period_days: int | None,
 ) -> list[dict[str, Any]]:
     """Prepare the periods to download a periodic mosaic.
 
@@ -365,7 +366,7 @@ def _prepare_periods(
 
 
 def _prepare_period_params(
-    period_name: Optional[str], period_days: Optional[int]
+    period_name: str | None, period_days: int | None
 ) -> tuple[str, int]:
     """Interprete period_name and period_days and return cleaned version.
 
@@ -392,19 +393,18 @@ def _prepare_period_params(
         period_days = 7
     elif period_name == "biweekly":
         period_days = 14
-    else:
-        if period_days is None:
-            raise ValueError(
-                "If period_name is not one of the basic names, period_days must be "
-                "specified."
-            )
+    elif period_days is None:
+        raise ValueError(
+            "If period_name is not one of the basic names, period_days must be "
+            "specified."
+        )
 
     return (period_name, period_days)
 
 
 def _prepare_periodic_mosaic_params(
     roi_bounds: tuple[float, float, float, float],
-    roi_crs: Optional[pyproj.CRS],
+    roi_crs: pyproj.CRS | None,
     start_date: datetime,
     end_date: datetime,
     imageprofiles_to_get: list[str],
@@ -505,11 +505,11 @@ def _prepare_periodic_mosaic_params(
 
 def _prepare_mosaic_image_params(
     roi_bounds: tuple[float, float, float, float],
-    roi_crs: Optional[pyproj.CRS],
+    roi_crs: pyproj.CRS | None,
     imageprofile: ImageProfile,
     period: dict[str, Any],
     output_base_dir: Path,
-    base_imageprofile: Optional[ImageProfile] = None,
+    base_imageprofile: ImageProfile | None = None,
 ) -> dict[str, Any]:
     """Prepares the relevant parameters + saves them as json file.
 
@@ -627,10 +627,10 @@ def _prepare_mosaic_image_path(
     return image_path
 
 
-def _imagemeta_to_file(path: Path, imagemeta: dict[str, Any]):
+def _imagemeta_to_file(path: Path, imagemeta: dict[str, Any]) -> None:
     # Write to file
     path.parent.mkdir(exist_ok=True, parents=True)
-    with open(path, "w") as outfile:
+    with path.open("w") as outfile:
         # Prepare some properties that don't serialize automatically
         imagemeta_prepared = deepcopy(imagemeta)
         imagemeta_prepared["start_date"] = imagemeta["start_date"].strftime("%Y-%m-%d")