+++ cropclassification/util/zonal_stats_bulk/__init__.py
@@ -1,8 +1,12 @@
 """Calculate zonal statistics for a vector file with many raster files."""
 
 from pathlib import Path
-from typing import Optional, Union
 
+from . import (
+    _zonal_stats_bulk_exactextract,
+    _zonal_stats_bulk_pyqgis,
+    _zonal_stats_bulk_rs,
+)
 from ._raster_helper import *  # noqa: F403
 
 
@@ -11,14 +15,14 @@ def zonal_stats(
     id_column: str,
     rasters_bands: list[tuple[Path, list[str]]],
     output_dir: Path,
-    stats: Union[list[str], str] = ["count", "median"],
-    cloud_filter_band: Optional[str] = None,
+    stats: list[str] | str | None = None,
+    cloud_filter_band: str | None = None,
     calc_bands_parallel: bool = True,
     engine: str = "rasterstats",
     nb_parallel: int = -1,
     force: bool = False,
-):
-    """Calculate zonal statistics.
+) -> None:
+    """Calculate zonal statistics and save the results in the output directory.
 
     Args:
         vector_path (Path): input file with vector data.
@@ -27,8 +31,9 @@ def zonal_stats(
         rasters_bands (List[Tuple[Path, List[str]]]): List of tuples with the path to
             the raster files and the bands to calculate the zonal statistics on.
         output_dir (Path): directory to write the results to.
-        stats (List[str]): statistics to calculate. Available statistics and
-            special options are dependent on the `engine` specified:
+        stats (List[str]): statistics to calculate. Default to ["count", "median"].
+            Available statistics and special options are dependent on the `engine`
+            specified:
 
                 - "rasterstats": `rasterstats documentation <https://pythonhosted.org/rasterstats/manual.html#statistics>`_
                 - "pyqgis": "count", "sum", "mean", "median", "std", "min", "max",
@@ -46,7 +51,9 @@ def zonal_stats(
         force (bool, optional): False to skip calculating existing output files. True to
             recalculate and overwrite existing output files. Defaults to False.
     """
-    if isinstance(stats, str):
+    if stats is None:
+        stats = ["count", "median"]
+    elif isinstance(stats, str):
         stats = [stats]
 
     if engine == "pyqgis":
@@ -54,7 +61,6 @@ def zonal_stats(
             raise ValueError(
                 'cloud_filter_band parameter not supported for engine "pyqgis"'
             )
-        from . import _zonal_stats_bulk_pyqgis
 
         return _zonal_stats_bulk_pyqgis.zonal_stats(
             vector_path=vector_path,
@@ -66,8 +72,6 @@ def zonal_stats(
             force=force,
         )
     elif engine == "rasterstats":
-        from . import _zonal_stats_bulk_rs
-
         return _zonal_stats_bulk_rs.zonal_stats(
             vector_path=vector_path,
             id_column=id_column,
@@ -80,8 +84,6 @@ def zonal_stats(
             force=force,
         )
     elif engine == "exactextract":
-        from . import _zonal_stats_bulk_exactextract
-
         return _zonal_stats_bulk_exactextract.zonal_stats(
             vector_path=vector_path,
             rasters_bands=rasters_bands,