+++ cropclassification/util/zonal_stats_bulk/_zonal_stats_bulk_pyqgis.py
@@ -8,7 +8,6 @@ import tempfile
 from concurrent import futures
 from datetime import datetime
 from pathlib import Path
-from typing import Union
 
 import geofileops as gfo
 import geopandas as gpd
@@ -51,7 +50,7 @@ def zonal_stats(
     stats: list[str],
     nb_parallel: int = -1,
     force: bool = False,
-):
+) -> None:
     """Calculate zonal statistics."""
     # Make sure QGIS is available
     if not HAS_QGIS:
@@ -77,7 +76,8 @@ def zonal_stats(
     # Loop over all images and bands to calculate zonal stats in parallel...
     calc_queue = {}
     pool = futures.ProcessPoolExecutor(
-        max_workers=nb_parallel, initializer=processing_util.initialize_worker()
+        max_workers=nb_parallel,
+        initializer=processing_util.initialize_worker(),  # type: ignore[func-returns-value]
     )
     try:
         for raster_path, bands in rasters_bands:
@@ -216,15 +216,15 @@ def zonal_stats(
 
 
 def zonal_stats_band(
-    vector_path,
+    vector_path: Path,
     raster_path: Path,
     band: str,
     tmp_dir: Path,
     stats: list[str],
     columns: list[str],
-) -> Union[pd.DataFrame, gpd.GeoDataFrame]:
+) -> pd.DataFrame | gpd.GeoDataFrame:
     # Init
-    stats_mask = None
+    stats_mask: int | None = None
     for stat in stats:
         if stats_mask is None:
             stats_mask = stat_to_qgisstat(stat)
@@ -285,12 +285,15 @@ def zonal_stats_band(
     if "geometry" in columns:
         columns = [f.name() for f in vector_mem.fields()] + ["geometry"]
         data = [
-            dict(zip(columns, [*f.attributes(), f.geometry().asWkt()]))
+            dict(zip(columns, [*f.attributes(), f.geometry().asWkt()], strict=True))
             for f in vector_mem.getFeatures()
         ]
     else:
         columns = [f.name() for f in vector_mem.fields()]
-        data = [dict(zip(columns, f.attributes())) for f in vector_mem.getFeatures()]
+        data = [
+            dict(zip(columns, f.attributes(), strict=True))
+            for f in vector_mem.getFeatures()
+        ]
     stats_df = pd.DataFrame(data, columns=columns)
 
     if "geometry" in columns:
@@ -303,7 +306,7 @@ def zonal_stats_band(
 
 
 def zonal_stats_band_tofile(
-    vector_path,
+    vector_path: Path,
     raster_path: Path,
     output_band_path: Path,
     band: str,
@@ -336,8 +339,8 @@ def zonal_stats_band_tofile(
     return output_band_path
 
 
-def stat_to_qgisstat(stat: str):
-    import qgis.analysis
+def stat_to_qgisstat(stat: str) -> int:
+    import qgis.analysis  # noqa: PLC0415
 
     if stat == "count":
         return qgis.analysis.QgsZonalStatistics.Count